Staging for DirectTrajOpt.Solvers minor structural refactor
