--- src/solvers.jl
@@ -1,10 +1,0 @@
-module Solvers
-
-export AbstractSolverOptions
-export solve!
-
-abstract type AbstractSolverOptions end
-
-function solve! end
-
-end
