+++ src/solvers/madnlp_solver/MadNLPSolverExt.jl
@@ -1,0 +1,19 @@
+module MadNLPSolverExt
+
+
+using DirectTrajOpt
+# using NamedTrajectories
+# using TrajectoryIndexingUtils
+
+import MathOptInterface as MOI
+import MadNLP
+
+using TestItemRunner
+
+
+using ..Solvers
+
+include("options.jl")
+include("solver.jl")
+
+end
